requirements.txt
python-dotenv>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0
python-srt>=3.5.0 